refactor: move draw_buffer sources into own directory
